requirements.txt
discord.py==2.3.1
requests==2.28.1
